Added more ship tests
Also:
Inject ShipComponentFactory into GameObjectFactory
Changed Ship.Damage to use floats instead of ints

+++ SpaceWar2/GameObjects/Ship.cs
@@ -53,7 +53,7 @@ namespace DEMW.SpaceWar2.GameObjects
             _thrusterArray.EngageThrusters();
         }
         
-        public void Damage(int amount)
+        public void Damage(float amount)
         {
             var damageRemaining = _shield.Damage(amount);
